rename: 공통컴포넌트 폴더구조 정리
