+++ src/components/dead-zone/DeadZone.tsx
