+++ src/components/dead-zone/DeadZonePopup.tsx
