+++ src/components/indicator/Indicator.tsx
